+++ client/src/App.tsx
@@ -3,7 +3,7 @@ import axios from 'axios'
 import { Router } from '@reach/router'
 
 import BaseLayout from './layouts/BaseLayout'
-import Index from './pages/Index'
+import Home from './pages/Home'
 import NotFound from './pages/NotFound'
 import Game from './pages/Game'
 
@@ -13,7 +13,7 @@ const App: React.FC = () => {
   return (
     <BaseLayout>
       <Router>
-        <Index path='/' />
+        <Home path='/' />
         <Game path='/game/:id' />
         <NotFound path='*' />
       </Router>
