+++ client/src/pages/Home/Home.scss
@@ -1,0 +1,5 @@
+@import 'styles/variables';
+
+.Home__GameList {
+  margin-top: $spacing-large;
+}
